cn/requirements.txt
chardet
baidu-aip
python-osc
SpeechRecognition
pyaudio